View

+++ src/shared_utils.c
@@ -1,28 +1,36 @@
+#define _XOPEN_SOURCE 500
+
 #include <shared_utils.h>
 #include <error_handler.h>
 #include <fcntl.h>
 #include <semaphore.h>
 #include <sys/stat.h>
 #include <sys/mman.h>
+#include <sys/types.h>
 #include <stddef.h>
 #include <unistd.h>
+#include <unistd.h>
+#include <errno.h>
 
 #define S_RWALL S_IWUSR | S_IRUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH
 
-#define INITIAL_VALUE 1
+
+
+#define INITIAL_VALUE 0
 
 void *open_shm(size_t shm_size) {
     int shm_fd = shm_open(SHM_NAME, O_CREAT | O_RDWR, S_RWALL);
     if(shm_fd == ERROR ) {
         exit_error("ERROR: Unable to open shared memory");
     }
-    void *shm = mmap(NULL, shm_size, PROT_READ | PROT_WRITE, MAP_SHARED, shm_fd, 0);
-    if(shm == MAP_FAILED) {
-        exit_error("ERROR: Unable to map shared memory");
-    }
 
     if (ftruncate(shm_fd, shm_size) < 0) {
         exit_error("ERROR: ftruncate");
+    }
+
+    void *shm = mmap(NULL, shm_size, PROT_READ | PROT_WRITE, MAP_SHARED, shm_fd, 0);
+    if(shm == MAP_FAILED) {
+        exit_error("ERROR: Unable to map shared memory");
     }
 
     if (close(shm_fd) == ERROR) {
@@ -46,7 +54,11 @@ void close_shm(void *shm_addr, char * name, size_t shm_size) {
     }
     //unlink shared memory
     if(shm_unlink(name) == ERROR) {
-        exit_error("ERROR: Unable to unlink shared memory");
+        if (errno != ENOENT) {
+            //Si es 2, la zona de memoria ya no existe
+            exit_error("ERROR: Unable to unlink shared memory");
+
+        }
     }
 }
